+++ src/plugins/plugin-player.js
@@ -13,8 +13,8 @@
  *
  * @param {function} initPlayerState - Function of type (playerID) => playerState.
  */
-export default initPlayerState => ({
-  wrapper: moveFn => {
+export default {
+  fnWrap: moveFn => {
     return (G, ctx, ...args) => {
       const current = ctx.currentPlayer;
       const player = G.players[current];
@@ -48,15 +48,17 @@ export default initPlayerState => ({
     };
   },
 
-  setup: (G, ctx) => {
-    let players = {};
-    for (let i = 0; i < ctx.numPlayers; i++) {
-      const playerState = {};
-      if (initPlayerState !== undefined) {
-        initPlayerState(i + '');
+  G: {
+    setup: (G, ctx, game) => {
+      let players = {};
+      for (let i = 0; i < ctx.numPlayers; i++) {
+        let playerState = {};
+        if (game.playerSetup !== undefined) {
+          playerState = game.playerSetup(i + '');
+        }
+        players[i + ''] = playerState;
       }
-      players[i + ''] = playerState;
-    }
-    return { ...G, players };
+      return { ...G, players };
+    },
   },
-});
+};